fixed services for prolly the last time

+++ src/main/java/com/example/hospitalapi/service/impl/HospitalServiceImpl.java
@@ -9,6 +9,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 import static com.example.hospitalapi.mapper.HospitalMapper.HOSPITAL_MAPPER;
 
@@ -21,6 +22,8 @@ public class HospitalServiceImpl implements HospitalService {
     private final RoomRepository roomRepository;
     private final DoctorRepository doctorRepository;
     private final BedRepository bedRepository;
+    private final BedRepository patientRepository;
+
 
     @Override
     public List<HospitalResource> findAll() {
@@ -41,18 +44,29 @@ public class HospitalServiceImpl implements HospitalService {
     }
 
     @Override
-    public HospitalResource findById(Long id) {
+    public Optional<HospitalResource> findById(Long id) {
         Hospital hospital = hospitalRepository.findById(id).get();
         hospital.setRooms(roomRepository.findAllByHospitalId(id));
         hospital.setAmbulances(ambulanceRepository.findAllByHospitalId(id));
         hospital.setDoctors(doctorRepository.findAllByHospitalId(id));
         hospital.setOperations(operationRepository.findAllByHospitalId(id));
 
-        return HOSPITAL_MAPPER.toHospitalResource(hospitalRepository.findById(id).get());
+        return Optional.of(HOSPITAL_MAPPER.toHospitalResource(hospital));
     }
 
     @Override
     public void deleteById(Long id) {
+        if(!hospitalRepository.existsById(id)) {
+            throw new RuntimeException("Hospital with id " + id + " does not exist");
+        }
+        ambulanceRepository.deleteAllByHospitalId(id);
+        operationRepository.deleteAllByHospitalId(id);
+        for(Room room : roomRepository.findAllByHospitalId(id)) {
+            patientRepository.deleteAllByRoomId(room.getId());
+            bedRepository.deleteAllByRoomId(room.getId());
+        }
+        roomRepository.deleteAllByHospitalId(id);
+        doctorRepository.deleteAllByHospitalId(id);
         hospitalRepository.deleteById(id);
     }
